requirements.txt
python-dotenv
flask
anthropic
google-generativeai
langchain
langchain-anthropic
langchain-openai
openai
PyPDF2
requests
werkzeug
uuid